CSV Bulk Question upload

+++ app/routes/$workspaceId/tests/$sectionId.tsx
@@ -13,7 +13,9 @@ import { routes } from "~/constants/route.constants"
 import { deleteQuestionStatus } from "~/interface/Interface"
 import { checkUserFeatureAuthorization } from "~/models/authorization.server"
 import {
+  addQuestionByCSV,
   deleteTestQuestionById,
+  getQuestionTypeFromTests,
   getSectionDataById,
 } from "~/services/tests.service"
 import { getUserId } from "~/session.server"
@@ -23,6 +25,7 @@ export const loader: LoaderFunction = async ({ request, params }) => {
   try {
     const userId = await getUserId(request)
     const currentWorkspaceId = params.workspaceId as string
+    const questionTypes = await getQuestionTypeFromTests()
     const permission = await checkUserFeatureAuthorization(
       userId!,
       currentWorkspaceId
@@ -39,7 +42,12 @@ export const loader: LoaderFunction = async ({ request, params }) => {
     if (!sectionDetails) {
       throw new Response("Not Found", { status: 404 })
     }
-    return json({ sectionDetails, currentWorkspaceId, permission })
+    return json({
+      sectionDetails,
+      currentWorkspaceId,
+      permission,
+      questionTypes,
+    })
   } catch (error: any) {
     if (error.status === 403) {
       return redirect(routes.unauthorized)
@@ -52,7 +60,22 @@ export const action: ActionFunction = async ({ request, params }) => {
     const id = formData.get("id") as string
     const userId = await getUserId(request)
     const currentWorkspaceId = params.workspaceId as string
-    return await deleteTestQuestionById(id, userId!, currentWorkspaceId)
+
+    const questions = formData.get("questionsData")
+    const action = formData.get("action")
+
+    if (action === "add-question-by-csv") {
+      await addQuestionByCSV(
+        JSON.parse(questions as any),
+        userId!,
+        currentWorkspaceId
+      )
+      return { msg: "success" }
+    }
+    if (action === "deleteQuestion") {
+      return await deleteTestQuestionById(id, userId!, currentWorkspaceId)
+    }
+    return { msg: "success" }
   } catch (error: any) {
     if (error.status === 403) {
       return redirect(routes.unauthorized)
